Refactor common canvas
