+++ canvas_modules/common-canvas/src/common-canvas/cc-central-items.jsx
@@ -17,6 +17,8 @@
 import React from "react";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
+import Palette from "../palette/palette.jsx";
+import CommonCanvasTooltip from "./cc-tooltip.jsx";
 import CanvasContents from "./cc-contents.jsx";
 import CommonCanvasToolbar from "./cc-toolbar.jsx";
 import CommonCanvasRightFlyout from "./cc-right-flyout.jsx";
@@ -31,23 +33,57 @@ class CommonCanvasCentralItems extends React.Component {
 		this.logger = new Logger("CC-CentralItems");
 	}
 
+	// Prevent the default behavior (which is to show a plus-sign pointer) as
+	// an object is being dragged over the common canvas components.
+	// Note: this is overriden by the canvas area itself to allow external objects
+	// to be dragged over it.
+	onDragOver(evt) {
+		evt.preventDefault();
+	}
+
+	// Prevent an object being dropped on the common canvas causing a file
+	// download event (which is the default!). Note: this is overriden by the
+	// canvas area itself to allow external objects to be dropped on it.
+	onDrop(evt) {
+		evt.preventDefault();
+	}
+
+	generateClass() {
+		let className = "common-canvas";
+
+		className += (
+			!this.props.enableEditingActions
+				? " config-editing-actions-false"
+				: "");
+
+		className += (
+			this.props.enableParentClass
+				? " " + this.props.enableParentClass
+				: "");
+
+		return className;
+	}
+
 	render() {
 		this.logger.log("render");
 
+		const tip = (<CommonCanvasTooltip canvasController={this.props.canvasController} />);
+		const palette = (<Palette canvasController={this.props.canvasController} containingDivId={this.props.containingDivId} />);
 		const rightFlyout = (<CommonCanvasRightFlyout />);
 		const canvasToolbar = (<CommonCanvasToolbar canvasController={this.props.canvasController} />);
-		const notificationPanel = (<NotificationPanel canvasController={this.props.canvasController} />);
 		const canvasContents = (<CanvasContents canvasController={this.props.canvasController} />);
 		const bottomPanel = (<CanvasBottomPanel canvasController={this.props.canvasController} containingDivId={this.props.containingDivId} />);
 		const topPanel = (<CanvasTopPanel canvasController={this.props.canvasController} containingDivId={this.props.containingDivId} />);
+		const notificationPanel = (<NotificationPanel canvasController={this.props.canvasController} />);
 
-		let centralItems = null;
+		let rightSideItems = null;
 		if (this.props.enableRightFlyoutUnderToolbar) {
 			const templateRows = this.props.toolbarIsOpen ? "auto 1fr" : "1fr";
 			const templateCols = this.props.rightFlyoutIsOpen ? "1fr auto" : "1fr";
 			let templateRows2 = this.props.topPanelIsOpen ? "auto 1fr" : "1fr";
 			templateRows2 += this.props.bottomPanelIsOpen ? " auto" : "";
-			centralItems = (
+
+			rightSideItems = (
 				<div className="common-canvas-right-side-items-under-toolbar" style={{ gridTemplateRows: templateRows }}>
 					{canvasToolbar}
 					<div id={this.props.containingDivId} className="common-canvas-items-container-under-toolbar" style={{ gridTemplateColumns: templateCols }}>
@@ -69,7 +105,7 @@ class CommonCanvasCentralItems extends React.Component {
 			templateRows += " 1fr"; // For canvas contents
 			templateRows += this.props.bottomPanelIsOpen ? " auto" : "";
 
-			centralItems = (
+			rightSideItems = (
 				<div className="common-canvas-right-side-items">
 					<div id={this.props.containingDivId} className="common-canvas-items-container" style={{ gridTemplateRows: templateRows }}>
 						{canvasToolbar}
@@ -83,7 +119,15 @@ class CommonCanvasCentralItems extends React.Component {
 			);
 		}
 
-		return centralItems;
+		const divClassName = this.generateClass();
+
+		return (
+			<div className={divClassName} onDragOver={this.onDragOver} onDrop={this.onDrop}>
+				{palette}
+				{rightSideItems}
+				{tip}
+			</div>
+		);
 	}
 }
 
@@ -93,6 +137,8 @@ CommonCanvasCentralItems.propTypes = {
 	containingDivId: PropTypes.string.isRequired,
 
 	// Provided by Redux
+	enableParentClass: PropTypes.string,
+	enableEditingActions: PropTypes.bool,
 	enableRightFlyoutUnderToolbar: PropTypes.bool,
 	toolbarIsOpen: PropTypes.bool,
 	topPanelIsOpen: PropTypes.bool,
@@ -101,6 +147,8 @@ CommonCanvasCentralItems.propTypes = {
 };
 
 const mapStateToProps = (state, ownProps) => ({
+	enableParentClass: state.canvasconfig.enableParentClass,
+	enableEditingActions: state.canvasconfig.enableEditingActions,
 	enableRightFlyoutUnderToolbar: state.canvasconfig.enableRightFlyoutUnderToolbar,
 	toolbarIsOpen: (state.canvasconfig.enableToolbarLayout !== "None"),
 	topPanelIsOpen: state.toppanel.isOpen,
